refactor(mcp): split http transport response format tests
