+++ packages/mcp/test/httpTransport.responseFormat.test.ts
@@ -2,54 +2,8 @@ import { describe, expect, it } from "vitest";
 
 import path from "node:path";
 
-import {
-  assertRecord,
-  parseFirstMcpPayload,
-  withMcpHttpServer,
-  withTempDir,
-} from "./httpTestUtils.js";
-
-const MCP_PROTOCOL_VERSION = "2025-03-26" as const;
-
-async function mcpInitializeRaw(options: {
-  url: string;
-  token: string;
-  clientName: string;
-  accept: string;
-}): Promise<{
-  status: number;
-  contentType: string;
-  sessionId: string | null;
-  body: string;
-  payload: unknown;
-}> {
-  const res = await fetch(options.url, {
-    method: "POST",
-    headers: {
-      Authorization: `Bearer ${options.token}`,
-      Accept: options.accept,
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({
-      jsonrpc: "2.0",
-      id: 1,
-      method: "initialize",
-      params: {
-        protocolVersion: MCP_PROTOCOL_VERSION,
-        capabilities: {},
-        clientInfo: { name: options.clientName, version: "0" },
-      },
-    }),
-  });
-
-  const status = res.status;
-  const contentType = res.headers.get("content-type") ?? "";
-  const sessionId = res.headers.get("mcp-session-id");
-
-  const body = await res.text();
-  const payload = parseFirstMcpPayload(body);
-  return { status, contentType, sessionId, body, payload };
-}
+import { assertRecord, withMcpHttpServer, withTempDir } from "./httpTestUtils.js";
+import { mcpInitializeRaw } from "./httpTransport.responseFormat.testUtils.js";
 
 describe("MCP HTTP server (Streamable HTTP response format)", () => {
   it("returns JSON by default when client accepts JSON", async () => {
@@ -111,145 +65,6 @@ describe("MCP HTTP server (Streamable HTTP response format)", () => {
         expect(res.sessionId).toBeTruthy();
         assertRecord(res.payload, "initialize payload");
         expect(res.payload).toHaveProperty("result");
-      });
-    });
-  });
-
-  it("accepts */* clients for GET SSE stream in JSON response mode (compat)", async () => {
-    await withTempDir("ailss-mcp-http-", async (dir) => {
-      const dbPath = path.join(dir, "index.sqlite");
-
-      await withMcpHttpServer({ dbPath, enableWriteTools: false }, async ({ url, token }) => {
-        const initRes = await mcpInitializeRaw({
-          url,
-          token,
-          clientName: "client-any-get-sse",
-          accept: "*/*",
-        });
-
-        expect(initRes.status).toBe(200);
-        expect(initRes.sessionId).toBeTruthy();
-
-        const res = await fetch(url, {
-          method: "GET",
-          headers: {
-            Authorization: `Bearer ${token}`,
-            Accept: "*/*",
-            "mcp-session-id": initRes.sessionId as string,
-          },
-        });
-
-        expect(res.status).toBe(200);
-        expect((res.headers.get("content-type") ?? "").startsWith("text/event-stream")).toBe(true);
-
-        if (res.body) {
-          await res.body.cancel();
-        }
-      });
-    });
-  });
-
-  it("does not coerce JSON-only clients for GET SSE stream in JSON response mode", async () => {
-    await withTempDir("ailss-mcp-http-", async (dir) => {
-      const dbPath = path.join(dir, "index.sqlite");
-
-      await withMcpHttpServer({ dbPath, enableWriteTools: false }, async ({ url, token }) => {
-        const initRes = await mcpInitializeRaw({
-          url,
-          token,
-          clientName: "client-json-only-get-sse",
-          accept: "*/*",
-        });
-
-        expect(initRes.status).toBe(200);
-        expect(initRes.sessionId).toBeTruthy();
-
-        const res = await fetch(url, {
-          method: "GET",
-          headers: {
-            Authorization: `Bearer ${token}`,
-            Accept: "application/json",
-            "mcp-session-id": initRes.sessionId as string,
-          },
-        });
-
-        expect(res.status).toBe(406);
-        expect((res.headers.get("content-type") ?? "").length > 0).toBe(true);
-
-        const body = await res.text();
-        const payload = parseFirstMcpPayload(body);
-        assertRecord(payload, "error payload");
-        expect(payload).toHaveProperty("error.message");
-      });
-    });
-  });
-
-  it("does not coerce GET clients that reject SSE via q-values", async () => {
-    await withTempDir("ailss-mcp-http-", async (dir) => {
-      const dbPath = path.join(dir, "index.sqlite");
-
-      await withMcpHttpServer({ dbPath, enableWriteTools: false }, async ({ url, token }) => {
-        const initRes = await mcpInitializeRaw({
-          url,
-          token,
-          clientName: "client-rejects-sse-get-sse",
-          accept: "*/*",
-        });
-
-        expect(initRes.status).toBe(200);
-        expect(initRes.sessionId).toBeTruthy();
-
-        const res = await fetch(url, {
-          method: "GET",
-          headers: {
-            Authorization: `Bearer ${token}`,
-            Accept: "text/*;q=0, application/json",
-            "mcp-session-id": initRes.sessionId as string,
-          },
-        });
-
-        expect(res.status).toBe(406);
-        expect((res.headers.get("content-type") ?? "").length > 0).toBe(true);
-
-        const body = await res.text();
-        const payload = parseFirstMcpPayload(body);
-        assertRecord(payload, "error payload");
-        expect(payload).toHaveProperty("error.message");
-      });
-    });
-  });
-
-  it("does not coerce parameterized wildcard Accept for GET SSE stream", async () => {
-    await withTempDir("ailss-mcp-http-", async (dir) => {
-      const dbPath = path.join(dir, "index.sqlite");
-
-      await withMcpHttpServer({ dbPath, enableWriteTools: false }, async ({ url, token }) => {
-        const initRes = await mcpInitializeRaw({
-          url,
-          token,
-          clientName: "client-param-wildcard-get-sse",
-          accept: "*/*",
-        });
-
-        expect(initRes.status).toBe(200);
-        expect(initRes.sessionId).toBeTruthy();
-
-        const res = await fetch(url, {
-          method: "GET",
-          headers: {
-            Authorization: `Bearer ${token}`,
-            Accept: "*/*;profile=v1",
-            "mcp-session-id": initRes.sessionId as string,
-          },
-        });
-
-        expect(res.status).toBe(406);
-        expect((res.headers.get("content-type") ?? "").length > 0).toBe(true);
-
-        const body = await res.text();
-        const payload = parseFirstMcpPayload(body);
-        assertRecord(payload, "error payload");
-        expect(payload).toHaveProperty("error.message");
       });
     });
   });
@@ -352,80 +167,6 @@ describe("MCP HTTP server (Streamable HTTP response format)", () => {
         expect(res.sessionId).toBeTruthy();
         assertRecord(res.payload, "initialize payload");
         expect(res.payload).toHaveProperty("result");
-      });
-    });
-  });
-
-  it("returns parseable JSON for notification-only POST bodies (compat)", async () => {
-    await withTempDir("ailss-mcp-http-", async (dir) => {
-      const dbPath = path.join(dir, "index.sqlite");
-
-      await withMcpHttpServer({ dbPath, enableWriteTools: false }, async ({ url, token }) => {
-        for (const accept of ["application/json", "text/event-stream"]) {
-          const initRes = await mcpInitializeRaw({
-            url,
-            token,
-            clientName: `client-notify-${accept.replace(/[^a-z0-9]+/gi, "-")}`,
-            accept,
-          });
-
-          expect(initRes.status).toBe(200);
-          expect(initRes.sessionId).toBeTruthy();
-
-          const res = await fetch(url, {
-            method: "POST",
-            headers: {
-              Authorization: `Bearer ${token}`,
-              Accept: accept,
-              "Content-Type": "application/json",
-              "Mcp-Session-Id": initRes.sessionId as string,
-              "Mcp-Protocol-Version": MCP_PROTOCOL_VERSION,
-            },
-            body: JSON.stringify({
-              jsonrpc: "2.0",
-              method: "notifications/initialized",
-              params: {},
-            }),
-          });
-
-          expect(res.status).toBe(202);
-          expect((res.headers.get("content-type") ?? "").startsWith("application/json")).toBe(true);
-          expect((await res.text()).trim()).toBe("null");
-        }
-
-        const batchInit = await mcpInitializeRaw({
-          url,
-          token,
-          clientName: "client-notify-batch",
-          accept: "application/json",
-        });
-
-        expect(batchInit.status).toBe(200);
-        expect(batchInit.sessionId).toBeTruthy();
-
-        const batchRes = await fetch(url, {
-          method: "POST",
-          headers: {
-            Authorization: `Bearer ${token}`,
-            Accept: "application/json",
-            "Content-Type": "application/json",
-            "Mcp-Session-Id": batchInit.sessionId as string,
-            "Mcp-Protocol-Version": MCP_PROTOCOL_VERSION,
-          },
-          body: JSON.stringify([
-            {
-              jsonrpc: "2.0",
-              method: "notifications/initialized",
-              params: {},
-            },
-          ]),
-        });
-
-        expect(batchRes.status).toBe(202);
-        expect((batchRes.headers.get("content-type") ?? "").startsWith("application/json")).toBe(
-          true,
-        );
-        expect((await batchRes.text()).trim()).toBe("[]");
       });
     });
   });
